+++ js/mock-names.js
@@ -1,0 +1,4 @@
+export const mockNamesExport = () => {
+  const mockNames = ['Матвей', 'Полина', 'Артём', 'Эмма', 'Есения', 'Денис', 'Андрей', 'Мирон', 'Владислав', 'Семён'];
+  return mockNames;
+};
